Clicking on tickets passes its informatin to the next page
